test(baes-picker): fix basepicker wrong test file name
